Title: 합과 평균, Time: 5ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 240801/합과 평균/sum-and-avg.cpp
@@ -6,11 +6,11 @@ int main() {
     int a, b;
 
     cin >> a >> b;
-    
+
     int c;
     c = a + b;
 
-    int d;
+    double d;
     d = (double) (a + b) / 2;
 
     cout << c << " " << d;
